Moment of great changes. Redesigned a lot of how the classes worked together.
Redesigned how the classes work, in particular the operand classes. Now a new MemoryStoredOperand abstract class has been added. Friend classes have been reduced to a minimum. Only BitralContext is friend to Operand classes , in order to be able to construct them.
The shared state of BitralContext with the Operand/CodeRegion classes is now encapsulated in a CompilerState class that is passed to the operand/coderegion costructor. The field itself inside BitralContext is private.